requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
pymongo==4.6.1
requests==2.31.0
PyJWT==2.8.0
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.1
